DEVPROD-750 remove legacy dispatcher versions
