+++ globals.go
@@ -278,8 +278,6 @@ const (
 	PlannerVersionLegacy  = "legacy"
 	PlannerVersionTunable = "tunable"
 
-	// TODO: EVG-18706 all distros use DispatcherVersionRevisedWithDependencies, we may be able to remove these and their custom logic
-	DispatcherVersionRevised                 = "revised"
 	DispatcherVersionRevisedWithDependencies = "revised-with-dependencies"
 
 	// maximum turnaround we want to maintain for all hosts for a given distro
@@ -958,7 +956,6 @@ var (
 
 	// Set of valid DispatchSettings.Version strings that can be user set via the API
 	ValidTaskDispatcherVersions = []string{
-		DispatcherVersionRevised,
 		DispatcherVersionRevisedWithDependencies,
 	}
