+++ rest/route/host_agent_test.go
@@ -911,388 +911,6 @@ func TestTaskLifecycleEndpoints(t *testing.T) {
 	}
 }
 
-func TestAssignNextAvailableTaskWithDispatcherSettingsVersionLegacy(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	env := &mock.Environment{}
-	require.NoError(t, env.Configure(ctx))
-
-	Convey("with a task queue and a host", t, func() {
-		settings := distro.DispatcherSettings{
-			Version: evergreen.DispatcherVersionRevisedWithDependencies,
-		}
-
-		colls := []string{distro.Collection, host.Collection, task.Collection, model.TaskQueuesCollection, model.ProjectRefCollection, build.Collection, model.VersionCollection}
-		require.NoError(t, db.ClearCollections(colls...))
-		defer func() {
-			assert.NoError(t, db.ClearCollections(colls...))
-		}()
-		require.NoError(t, modelUtil.AddTestIndexes(host.Collection, true, true, host.RunningTaskKey))
-		distroID := "testDistro"
-		d := distro.Distro{
-			Id:                 distroID,
-			DispatcherSettings: settings,
-		}
-		So(d.Insert(ctx), ShouldBeNil)
-
-		taskGroupInfo := model.TaskGroupInfo{
-			Name:  "",
-			Count: 2,
-		}
-		distroQueueInfo := model.DistroQueueInfo{
-			Length:         2,
-			TaskGroupInfos: []model.TaskGroupInfo{taskGroupInfo},
-		}
-		taskQueue := &model.TaskQueue{
-			Distro: distroID,
-			Queue: []model.TaskQueueItem{
-				{Id: "task1"},
-				{Id: "task2"},
-			},
-			DistroQueueInfo: distroQueueInfo,
-		}
-		So(taskQueue.Save(), ShouldBeNil)
-
-		theHostWhoCanBoastTheMostRoast := host.Host{
-			Id: "h1",
-			Distro: distro.Distro{
-				Id:                 distroID,
-				DispatcherSettings: settings,
-			},
-			Secret: hostSecret,
-			Status: evergreen.HostRunning,
-		}
-		So(theHostWhoCanBoastTheMostRoast.Insert(ctx), ShouldBeNil)
-
-		task1 := task.Task{
-			Id:        "task1",
-			Status:    evergreen.TaskUndispatched,
-			Activated: true,
-			Project:   "exists",
-		}
-		task2 := task.Task{
-			Id:        "task2",
-			Status:    evergreen.TaskUndispatched,
-			Activated: true,
-			Project:   "exists",
-		}
-		pref := &model.ProjectRef{
-			Id:      "exists",
-			Enabled: true,
-		}
-		So(task1.Insert(), ShouldBeNil)
-		So(task2.Insert(), ShouldBeNil)
-		So(pref.Insert(), ShouldBeNil)
-
-		details := &apimodels.GetNextTaskDetails{}
-
-		Convey("a host should get the task at the top of the queue", func() {
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(t, ShouldNotBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-			So(t.Id, ShouldEqual, "task1")
-
-			currentTq, err := model.LoadTaskQueue(distroID)
-			So(err, ShouldBeNil)
-			So(currentTq.Length(), ShouldEqual, 1)
-
-			h, err := host.FindOne(ctx, host.ById(theHostWhoCanBoastTheMostRoast.Id))
-			So(err, ShouldBeNil)
-			So(h.RunningTask, ShouldEqual, "task1")
-
-		})
-		Convey("tasks with a disabled project should be removed from the queue", func() {
-			pref.Enabled = false
-			So(pref.Upsert(), ShouldBeNil)
-
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(t, ShouldBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-
-			currentTq, err := model.LoadTaskQueue(distroID)
-			So(err, ShouldBeNil)
-			So(currentTq.Length(), ShouldEqual, 0)
-		})
-		Convey("tasks belonging to a project with dispatching disabled should be removed from the queue", func() {
-			pref.DispatchingDisabled = utility.TruePtr()
-			So(pref.Upsert(), ShouldBeNil)
-
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(t, ShouldBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-
-			currentTq, err := model.LoadTaskQueue(distroID)
-			So(err, ShouldBeNil)
-			So(currentTq.Length(), ShouldEqual, 0)
-		})
-		Convey("a completed task group should return a nil task", func() {
-			currentTq, err := model.LoadTaskQueue(distroID)
-			So(err, ShouldBeNil)
-			So(currentTq.Length(), ShouldEqual, 2)
-
-			details.TaskGroup = "my-task-group"
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchAliasService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(t, ShouldBeNil)
-			So(shouldTeardown, ShouldBeTrue)
-
-			// task queue unmodified
-			currentTq, err = model.LoadTaskQueue(distroID)
-			So(err, ShouldBeNil)
-			So(currentTq.Length(), ShouldEqual, 2)
-		})
-		Convey("a task that is not undispatched should not be updated in the host", func() {
-			taskQueue.Queue = []model.TaskQueueItem{
-				{Id: "undispatchedTask"},
-				{Id: "task2"},
-			}
-			So(taskQueue.Save(), ShouldBeNil)
-			undispatchedTask := task.Task{
-				Id:     "undispatchedTask",
-				Status: evergreen.TaskStarted,
-			}
-			So(undispatchedTask.Insert(), ShouldBeNil)
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-			So(t.Id, ShouldEqual, "task2")
-
-			currentTq, err := model.LoadTaskQueue(distroID)
-			So(err, ShouldBeNil)
-			So(currentTq.Length(), ShouldEqual, 0)
-		})
-		Convey("an empty task queue should return a nil task", func() {
-			taskQueue.Queue = []model.TaskQueueItem{}
-			So(taskQueue.Save(), ShouldBeNil)
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-			So(t, ShouldBeNil)
-		})
-		Convey("a tasks queue with a task that does not exist should continue", func() {
-			taskQueue.Queue = []model.TaskQueueItem{{Id: "notatask"}}
-			So(taskQueue.Save(), ShouldBeNil)
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-			So(t, ShouldBeNil)
-		})
-		Convey("a task's NumNextTaskDispatches should be increased by 1", func() {
-			taskQueue.Queue = []model.TaskQueueItem{{Id: "stillOkayTask"}}
-			So(taskQueue.Save(), ShouldBeNil)
-			stillOkayTask := task.Task{
-				Id:                    "stillOkayTask",
-				Status:                evergreen.TaskUndispatched,
-				Activated:             true,
-				NumNextTaskDispatches: 4,
-				BuildId:               "b1",
-				Version:               versionId,
-				HostId:                theHostWhoCanBoastTheMostRoast.Id,
-			}
-			So(stillOkayTask.Insert(), ShouldBeNil)
-			testBuild := build.Build{
-				Id:      "b1",
-				Version: versionId,
-			}
-			require.NoError(t, testBuild.Insert())
-
-			testVersion := model.Version{
-				Id: versionId,
-			}
-			require.NoError(t, testVersion.Insert())
-			t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &theHostWhoCanBoastTheMostRoast, details)
-			So(err, ShouldBeNil)
-			So(shouldTeardown, ShouldBeFalse)
-			So(t, ShouldNotBeNil)
-			taskBFromDb, err := task.FindOneId(stillOkayTask.Id)
-			So(err, ShouldBeNil)
-			So(taskBFromDb, ShouldNotBeNil)
-			So(taskBFromDb.IsStuckTask(), ShouldBeTrue)
-		})
-		Convey("with a host with a running task", func() {
-			anotherHost := host.Host{
-				Id:          "ahost",
-				RunningTask: "sampleTask",
-				Distro: distro.Distro{
-					Id: distroID,
-				},
-				Secret: hostSecret,
-			}
-			So(anotherHost.Insert(ctx), ShouldBeNil)
-			h2 := host.Host{
-				Id: "host2",
-				Distro: distro.Distro{
-					Id: distroID,
-				},
-				Secret: hostSecret,
-				Status: evergreen.HostRunning,
-			}
-			So(h2.Insert(ctx), ShouldBeNil)
-
-			t1 := task.Task{
-				Id:        "sampleTask",
-				Status:    evergreen.TaskUndispatched,
-				Project:   "exists",
-				Activated: true,
-			}
-			So(t1.Insert(), ShouldBeNil)
-			t2 := task.Task{
-				Id:        "another",
-				Status:    evergreen.TaskUndispatched,
-				Project:   "exists",
-				Activated: true,
-			}
-			So(t2.Insert(), ShouldBeNil)
-
-			taskQueue.Queue = []model.TaskQueueItem{
-				{Id: t1.Id},
-				{Id: t2.Id},
-			}
-			So(taskQueue.Save(), ShouldBeNil)
-			Convey("the task that is in the other host should not be assigned to another host", func() {
-				t, shouldTeardown, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &h2, details)
-				So(err, ShouldBeNil)
-				So(shouldTeardown, ShouldBeFalse)
-				So(t, ShouldNotBeNil)
-				So(t.Id, ShouldEqual, t2.Id)
-				h, err := host.FindOne(ctx, host.ById(h2.Id))
-				So(err, ShouldBeNil)
-				So(h.RunningTask, ShouldEqual, t2.Id)
-			})
-			Convey("a host with a running task should return an error", func() {
-				_, _, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &anotherHost, details)
-				So(err, ShouldNotBeNil)
-			})
-		})
-		Convey("with a host running a task in a task group", func() {
-			host1 := host.Host{
-				Id:                      "host1",
-				Status:                  evergreen.HostRunning,
-				RunningTask:             "task1",
-				RunningTaskGroup:        "group1",
-				RunningTaskBuildVariant: "variant1",
-				RunningTaskVersion:      "version1",
-				RunningTaskProject:      "exists",
-			}
-			So(host1.Insert(ctx), ShouldBeNil)
-			host2 := host.Host{
-				Id:                      "host2",
-				Status:                  evergreen.HostRunning,
-				RunningTask:             "",
-				RunningTaskGroup:        "",
-				RunningTaskBuildVariant: "",
-				RunningTaskVersion:      "",
-				RunningTaskProject:      "",
-			}
-			So(host2.Insert(ctx), ShouldBeNil)
-			task3 := task.Task{
-				Id:                "task3",
-				Status:            evergreen.TaskUndispatched,
-				Activated:         true,
-				TaskGroup:         "group1",
-				BuildVariant:      "variant1",
-				Version:           "version1",
-				Project:           "exists",
-				TaskGroupMaxHosts: 1,
-			}
-			So(task3.Insert(), ShouldBeNil)
-			task4 := task.Task{
-				Id:                "task4",
-				Status:            evergreen.TaskUndispatched,
-				Activated:         true,
-				TaskGroup:         "group2",
-				BuildVariant:      "variant1",
-				Version:           "version1",
-				Project:           "exists",
-				TaskGroupMaxHosts: 1,
-			}
-			So(task4.Insert(), ShouldBeNil)
-			taskQueue.Queue = []model.TaskQueueItem{
-				{Id: task3.Id},
-				{Id: task4.Id},
-			}
-			So(taskQueue.Save(), ShouldBeNil)
-			t, _, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &host2, details)
-			So(err, ShouldBeNil)
-			So(t, ShouldNotBeNil)
-			// task 3 should not be dispatched, because it's already running on max
-			// hosts, instead it should be task 4
-			So(t.Id, ShouldEqual, task4.Id)
-			h, err := host.FindOne(ctx, host.ById(host2.Id))
-			So(err, ShouldBeNil)
-			So(h.RunningTask, ShouldEqual, task4.Id)
-		})
-		Convey("with many hosts running task group tasks", func() {
-			// In this scenario likely host1 and host2 are racing, since host2 has a later
-			// task group order number than what's in the queue, and will clear the running
-			// task when it sees that host2 is running with a smaller task group order number.
-			host1 := host.Host{
-				Id:                      "host1",
-				Status:                  evergreen.HostRunning,
-				RunningTask:             "task1",
-				RunningTaskGroup:        "group1",
-				RunningTaskBuildVariant: "variant1",
-				RunningTaskVersion:      "version1",
-				RunningTaskProject:      "exists",
-				RunningTaskGroupOrder:   2,
-			}
-			So(host1.Insert(ctx), ShouldBeNil)
-			host2 := host.Host{
-				Id:                      "host2",
-				Status:                  evergreen.HostRunning,
-				RunningTask:             "",
-				RunningTaskGroup:        "",
-				RunningTaskBuildVariant: "",
-				RunningTaskVersion:      "",
-				RunningTaskProject:      "",
-			}
-			So(host2.Insert(ctx), ShouldBeNil)
-			task3 := task.Task{
-				Id:                "task3",
-				Status:            evergreen.TaskUndispatched,
-				Activated:         true,
-				TaskGroup:         "group1",
-				BuildVariant:      "variant1",
-				Version:           "version1",
-				Project:           "exists",
-				TaskGroupMaxHosts: 1,
-				TaskGroupOrder:    3,
-			}
-			So(task3.Insert(), ShouldBeNil)
-			task4 := task.Task{
-				Id:                "task4",
-				Status:            evergreen.TaskUndispatched,
-				Activated:         true,
-				TaskGroup:         "group1",
-				BuildVariant:      "variant1",
-				Version:           "version1",
-				Project:           "exists",
-				TaskGroupMaxHosts: 1,
-				TaskGroupOrder:    1,
-			}
-			So(task4.Insert(), ShouldBeNil)
-			taskQueue.Queue = []model.TaskQueueItem{
-				{Id: task3.Id},
-				{Id: task4.Id},
-			}
-			So(taskQueue.Save(), ShouldBeNil)
-			t, _, err := assignNextAvailableTask(ctx, env, taskQueue, model.NewTaskDispatchService(time.Minute), &host2, details)
-			So(err, ShouldBeNil)
-			So(t, ShouldNotBeNil)
-			// task 3 should not be dispatched, because it has a later task group
-			// order number than what's currently assigned to host1. Instead it should be task4.
-			So(t.Id, ShouldEqual, task4.Id)
-			h, err := host.FindOne(ctx, host.ById(host2.Id))
-			So(err, ShouldBeNil)
-			So(h.RunningTask, ShouldEqual, task4.Id)
-		})
-	})
-}
-
 func TestAssignNextAvailableTaskWithDispatcherSettingsVersionTunable(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
